# presentation/swagger/swagger.yaml
definitions:
...
  model.Armor:
    properties:
      alias:
        type: string
      arcaneSpellFailureChance:
        type: integer
      armorBonus:
        type: integer
      armorCheckPenalty:
        type: integer
      book:
        $ref: '#/definitions/model.Books'
      cost:
        type: number
      description:
        type: string
      engName:
        type: string
      maxDexBonus:
        type: integer
      name:
        type: string
      speed20:
        type: integer
      speed30:
        type: integer
      type:
        $ref: '#/definitions/model.NameAlias'
      weight:
        type: number
    type: object
...
  model.Books:
    properties:
      abbreviation:
        type: string
      alias:
        type: string
      name:
        type: string
      order:
        type: integer
    type: object
...
  model.ErrorResponse:
    properties:
      message:
        type: string
      title:
        type: string
    type: object
...
  model.GoodAndService:
    properties:
      alias:
        type: string
      book:
        $ref: '#/definitions/model.Books'
      childs:
        items:
          $ref: '#/definitions/model.GoodAndService'
        type: array
      cost:
        type: number
      costDescription:
        type: string
      costOfPassage:
        type: number
      costSpecial:
        type: string
      craftDc:
        type: integer
      description:
        type: string
      engName:
        type: string
      equipmentSubType:
        type: string
      name:
        type: string
      parents:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/model.NameAliasDescription'
      weight:
        type: number
      weightDescription:
        type: string
      weightSpecial:
        type: string
    type: object
...
  model.HelperStat:
    properties:
      alias:
        type: string
      cnt:
        type: integer
      name:
        type: string
    type: object
...
  model.NameAlias:
    properties:
      alias:
        type: string
      name:
        type: string
    type: object
  model.NameAliasDescription:
    properties:
      alias:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
...
  model.Thanks:
    properties:
      helperStat:
        items:
          $ref: '#/definitions/model.HelperStat'
        type: array
      type:
        type: string
    type: object
  model.Trait:
    properties:
      alias:
        type: string
      benefit:
        type: string
      book:
        $ref: '#/definitions/model.Books'
      engName:
        type: string
      name:
        type: string
      prerequisites:
        type: string
      type:
        $ref: '#/definitions/model.TraitType'
    type: object
  model.TraitType:
    properties:
      alias:
        type: string
      description:
        type: string
      name:
        type: string
      parentType:
        $ref: '#/definitions/model.TraitType'
    type: object
...
  model.Weapon:
    properties:
      alias:
        type: string
      book:
        $ref: '#/definitions/model.Books'
      capacity:
        type: integer
      childs:
        items:
          $ref: '#/definitions/model.Weapon'
        type: array
      cost:
        type: number
      criticalDamage:
        type: string
      criticalRoll:
        type: string
      damageM:
        type: string
      damageS:
        type: string
      description:
        type: string
      encumbranceCategory:
        $ref: '#/definitions/model.NameAlias'
      engName:
        type: string
      misfire:
        type: string
      name:
        type: string
      parents:
        items:
          type: string
        type: array
      proficientCategory:
        $ref: '#/definitions/model.NameAlias'
      range:
        type: string
      rangeCategory:
        $ref: '#/definitions/model.NameAlias'
      special:
        type: string
      type:
        type: string
      weight:
        type: number
    type: object
...
info:
  contact: {}
  description: Сайт с информацией для настольной игры Pathfinder
  license:
    name: Чудинов Андрей Дмитриевич
  title: Pathfinder Family
  version: "20220823"
paths:
...
  /api/armors:
    get:
      description: Список доспехов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Armor'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список доспехов
      tags:
      - API
...
  /api/goodsAndServices:
    get:
      description: Список товаров и услуг
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.GoodAndService'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список товаров и услуг
      tags:
      - API
...
  /api/thanks:
    get:
      description: Получение всех помощников по группам
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Thanks'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Получение всех помощников по группам
      tags:
      - API
  /api/traits:
    get:
      description: Список штрихов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Trait'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список штрихов
      tags:
      - API
...
  /api/weapons:
    get:
      description: Список оружия
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Weapon'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список оружия
      tags:
      - API
...
  /armors:
    get:
      description: Доспехи
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Доспехи
      tags:
      - Site
  /armors/description:
    get:
      description: Общее описание доспехов
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Общее описание доспехов
      tags:
      - Site
...
  /class/{alias}/spirit:
    get:
      description: Духи
      parameters:
      - description: Alias класса
        in: path
        name: alias
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Духи
      tags:
      - Site
...
  /goodsAndServices:
    get:
      description: Товары и услуги
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Товары и услуги
      tags:
      - Site
...
  /skill:
    get:
      description: Навыки
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Навыки
      tags:
      - Site
  /skill/{alias}:
    get:
      description: Информация по навыку
      parameters:
      - description: Alias навыка
        in: path
        name: alias
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Информация по навыку
      tags:
      - Site
...
  /specialMaterials:
    get:
      description: Особые материалы
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Особые материалы
      tags:
      - Site
...
  /thanks:
    get:
      description: Благодарности
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Благодарности
      tags:
      - Site
...
  /traits:
    get:
      description: Штрихи персонажей
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Штрихи персонажей
      tags:
      - Site
...
  /wealthAndMoney:
    get:
      description: Богатство и деньги
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Богатство и деньги
      tags:
      - Site
  /weapons:
    get:
      description: Оружие
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Оружие
      tags:
      - Site
  /weapons/description:
    get:
      description: Общее описание оружия
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Общее описание оружия
      tags:
      - Site
...
swagger: "2.0"